Tabular: Moved, renamed, and optimized generate_text_features
